+++ gazon-service/src/main/java/com/gitlab/service/PassportService.java
@@ -18,7 +18,7 @@ import java.util.Optional;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class PassportService implements Cloneable{
+public class PassportService implements Cloneable {
 
     private final PassportRepository passportRepository;