+++ gazon-service/src/main/java/com/gitlab/service/PersonalAddressService.java
@@ -18,7 +18,7 @@ import java.util.Optional;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class PersonalAddressService implements Cloneable{
+public class PersonalAddressService implements Cloneable {
 
     private final PersonalAddressRepository personalAddressRepository;