+++ gazon-service/src/test/java/com/gitlab/dto/PaymentDtoTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -58,11 +59,13 @@ class PaymentDtoTest extends AbstractDtoTest {
     }
 
     @Test
-    void test_invalid_userId(){
+    void test_invalid_shouldSaveCard() {
         var paymentDto = generatePaymentDto();
-        paymentDto.setUserId(null);
-
-        assertFalse(validator.validate(paymentDto).isEmpty());
+        assertTrue(validator.validate(paymentDto).isEmpty());
+        paymentDto.setShouldSaveCard(true);
+        assertThat(paymentDto.isShouldSaveCard()).isTrue();
+        paymentDto.setShouldSaveCard(false);
+        assertThat(paymentDto.isShouldSaveCard()).isFalse();
     }
 
     private PaymentDto generatePaymentDto() {
@@ -76,7 +79,6 @@ class PaymentDtoTest extends AbstractDtoTest {
 
         paymentDto.setOrderId(1L);
         paymentDto.setSum(new BigDecimal(500));
-        paymentDto.setUserId(1L);
         return paymentDto;
     }
 }